@alekseykulikov 2016-07-04
I'd like to setup jsperf to make sure it works fast in other engines, but...
![screen shot 2016-07-04 at 14 00 50](https://cloud.githubusercontent.com/assets/158189/16559828/e17a58b4-41ef-11e6-93b7-315d89441acc.png)