@dcousens 2016-07-04
> How: use string intead of array

https://github.com/JedWatson/classnames/pull/50
https://github.com/JedWatson/classnames/pull/36